Refactor API graph primitives
